Add various caption box enhancements.
Closes #503.
