+++ lib/control_panel/dialogs/advanced_controls_dialog.ts
@@ -61,12 +61,12 @@ export class AdvancedControlSettingsDialog extends logging(ParaComponent) {
     this._store.settingControls.add({
       type: 'checkbox',
       key: 'controlPanel.isStatusBarVisible',
-      label: 'Status bar visible',
+      label: 'Exploration bar visible',
       parentView: 'controlPanel.tabs.controls.dialog.settings',
     });
     this._store.settingControls.add({
       type: 'checkbox',
-      key: 'controlPanel.isCaptionExternalWhenControlPanelClosed',
+      key: 'controlPanel.caption.isCaptionExternalWhenControlPanelClosed',
       label: 'Caption external when control panel closed',
       parentView: 'controlPanel.tabs.controls.dialog.settings',
     });
