+++ lib/parachart/style_manager.ts
